requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
# llama-cpp-python - install separately with precompiled wheels
diskcache>=5.6.1
jinja2>=2.11.3
numpy>=1.20.0
typing-extensions>=4.5.0